LBL Web Challenge
This is a webpage that I created from scratch using some basic JS
principles, a new JS framework in chart.js, and a JS plugin called
sorttable.js for a web challenge for a job application with the
Lawrence Berkeley Labs. It’s very much at the structure level and there
is more design that can be done, but I though that I’d put it up here
to show some of my JS knowledge and that I’m learning to work with
other JS frameworks and their documentation.